requirements.txt
Flask==2.2.5
google.auth==2.41.1
requests==2.32.5
google-cloud-asset==3.30.1
protobuf==6.32.1
google-api-python-client==2.186.0